Impl Niching for Null without alloc feature
